pyproject.toml
[project]
name = "music-sync"
version = "0.1.0"
description = "Sync your downloaded music with your online platforms"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorlog>=6.9.0",
    "marshmallow>=3.26.0",
    "pyxdg>=0.28",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "yt-dlp>=2025.1.15",
]

[dependency-groups]
dev = [
    "ruff>=0.9.3",
]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true